Move things from message, update dependencies
