+++ src/app/util/types_stub.h
@@ -20,9 +20,9 @@
 #include <string.h> // For mem* functions.
 
 #include <app/util/basic-types.h>
+#include <lib/core/NodeId.h>
 #include <lib/core/Optional.h>
 
-#include <transport/raw/MessageHeader.h>
 static_assert(sizeof(chip::NodeId) == sizeof(uint64_t), "Unexpected node if size");
 
 /**
