+++ src/app/util/util.cpp
@@ -25,6 +25,7 @@
 #include <app/util/attribute-storage.h>
 #include <app/util/config.h>
 #include <app/util/generic-callbacks.h>
+#include <app/util/strings.h>
 #include <lib/core/CHIPConfig.h>
 #include <lib/core/CHIPEncoding.h>
 #include <protocols/interaction_model/StatusCode.h>
