+++ src/app/zap-templates/templates/app/attributes/Accessors.zapt
@@ -9,6 +9,7 @@
 
 #include <app/data-model/Nullable.h>
 #include <app/util/af-types.h>
+#include <app/util/strings.h>
 #include <app-common/zap-generated/cluster-objects.h>
 #include <lib/support/Span.h>
 #include <protocols/interaction_model/StatusCode.h>
