+++ zzz_generated/app-common/app-common/zap-generated/attributes/Accessors.h
@@ -27,6 +27,7 @@
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app/data-model/Nullable.h>
 #include <app/util/af-types.h>
+#include <app/util/strings.h>
 #include <lib/support/Span.h>
 #include <protocols/interaction_model/StatusCode.h>
